moved validated structs to controller/mod.rs, renamed controller -> trino_controller, pass through validated cluster
